+++ inst/process_zoo.R
@@ -1,7 +1,4 @@
 extract_genus_species <- function(taxa_vector) {
-  require(stringr)
-  require(dplyr)
-
   tibble::tibble(
     original_name = taxa_vector,
     genus_species = case_when(
@@ -47,16 +44,16 @@ extract_genus_species <- function(taxa_vector) {
     )
   )
 }
-usethis::use_devtools()
 
 dat <- readxl::read_xlsx(
-  "lter_zoo_84_13.xlsx",
+  "data/lter_zoo_84_13.xlsx",
   .name_repair = "minimal"
-)
+) |>
+  dplyr::mutate(dat_id = seq_len(dplyr::n()))
 
 ids <-
   readxl::read_xlsx(
-    "ids.xlsx",
+    "data/ids.xlsx",
     .name_repair = "minimal"
   ) |>
   dplyr::mutate(
@@ -64,21 +61,60 @@ ids <-
     sample_id = janitor::make_clean_names(sample_id)
   )
 
-
-tidy_data <-
+dates <-
   dat |>
-  dplyr::select(-c(1:7, "NOTES")) |>
+  dplyr::select(-c(1:10), "dat_id")
+
+
+raw_taxa <-
+  dat |>
+  janitor::clean_names() |>
+  dplyr::select("phylum_o_subphylum":"stage", "dat_id") |>
+  dplyr::mutate(
+    genus_species = extract_genus_species(.data$taxa)$genus_species,
+  ) |>
+  tidyr::separate(
+    genus_species,
+    into = c("genus", "species"),
+    sep = " ",
+    remove = FALSE
+  ) |>
+  dplyr::mutate(taxa_worrms = paste(.data$genus_subgenus, .data$species))
+
+worrms_matched <-
+  raw_taxa |>
+  dplyr::select("dat_id", "taxa_worrms") |>
+  dplyr::rowwise() |>
+  dplyr::mutate(
+    worrms_match = tryCatch(
+      {
+        worrms::wm_records_taxamatch(name = taxa_worrms, verbose = FALSE)
+      },
+      error = function(e) {
+        # Return NA or empty tibble when there's an error
+        NA
+      }
+    )
+  ) |>
+  dplyr::select(-"taxa_worrms")
+
+
+raw_taxa |>
+  dplyr::select(-"taxa_worrms") |>
+  dplyr::left_join(worrms_matched, by = "dat_id") |>
+  dplyr::select("dat_id", "taxa", "worrms_match", "stage") |>
+  dplyr::left_join(dates, by = "dat_id") |>
   tidyr::pivot_longer(
-    -c("TAXA", "stage"),
+    -c("dat_id", "taxa", "worrms_match", "stage"),
     names_to = "date",
     values_to = "ind_m3"
   ) |>
   janitor::clean_names() |>
   dplyr::mutate(
     date = lubridate::as_date(as.numeric(date), origin = "1899-12-30"),
-    #taxa_std = extract_genus_species(taxa)$genus_species
   ) |>
-  dplyr::full_join(ids, by = c("date")) |>
+  dplyr::select(-"dat_id") |>
+  dplyr::left_join(ids, by = c("date")) |>
   dplyr::relocate(sample_id, .before = "taxa") |>
   dplyr::relocate(stage, .after = "ind_m3") |>
   dplyr::relocate(date, .after = "sample_id") |>
@@ -127,8 +163,6 @@ occurrence_table <-
     lifeStage,
     occurrenceStatus
   )
-occurrence_table |>
-  dplyr::filter(scientificName == "Larvae n.i." & occurrenceStatus == "present")
 
 emof_table <-
   full_table |>
@@ -140,26 +174,3 @@ emof_table <-
     names_to = "meameasurementType",
     values_to = "measurementValue"
   )
-
-
-emof_table |>
-  dplyr::filter(occurrenceID == "mc_318-occ130001")
-
-
-occurrence_table |>
-  dplyr::filter(occurrenceID == "mc_1-occ29")
-
-
-full_table |>
-  dplyr::filter(
-    scientificName == "Chiridius poppei Giesbrecht, 1893",
-    eventID == "mc_1"
-  )
-
-
-we <-
-  full_table |>
-  dplyr::filter(individualCount > 0 & is.na(lifeStage))
-
-
-unique(we$scientificName)